clase_hoy/requirements.txt
streamlit 
pandas 
plotly 
pillow
pyzbar